gnomeExtension: Reformat extension.json
The idea of having one line per extension for diff reasons mostly worked out well.
However, especially around major Gnome updates it happens that extensions only change
in the set of supported versions. This information is currently a bit hard to track,
since such a change would still result in a diff for the entire line, which can be pretty long.
Additionally, it may happen that this applies to many extensions at once, and the way git
renders diff (first all removals in a hunk, then all additions) makes it even harder
to track because now the change is not local in the diff.

Having one line per supported version per extension should solve this problem, while
still providing a good compromise in overall line count of the file.
It shall be noted that GitHub already refuses to render diffs on this file even for
comparatively small updates, therefore this change cannot really make it worse here.